+++ backend/server/src/main/java/shop/sellution/server/order/presentation/OrderController.java
@@ -12,6 +12,7 @@ import shop.sellution.server.order.dto.OrderSearchCondition;
 import shop.sellution.server.order.dto.request.CancelOrderReq;
 import shop.sellution.server.order.dto.request.SaveOrderReq;
 import shop.sellution.server.order.dto.response.FindOrderRes;
+import shop.sellution.server.sms.application.SmsServiceImpl;
 
 @RestController
 @Slf4j
