+++ backend/server/src/main/java/shop/sellution/server/payment/application/PaymentHistoryServiceImpl.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import shop.sellution.server.payment.domain.repository.PaymentHistoryRepository;
 import shop.sellution.server.payment.dto.request.FindPaymentHistoryCond;
 import shop.sellution.server.payment.dto.response.FindPaymentHistoryRes;
@@ -15,6 +16,7 @@ public class PaymentHistoryServiceImpl implements PaymentHistoryService {
     private final PaymentHistoryRepository paymentHistoryRepository;
 
     @Override
+    @Transactional(readOnly = true)
     public Page<FindPaymentHistoryRes> findPaymentHistoryByCompanyId(Long companyId, FindPaymentHistoryCond findPaymentHistoryCond, Pageable pageable) {
         return paymentHistoryRepository.findPaymentHistoryByConditions(companyId, findPaymentHistoryCond, pageable);
     }
